Refactor cgroup logic into new file
No functional changes
